Move graphqlMocking file
